Add files via upload
Wrote in some of the generators. With that said, my syntax is starting to struggle. Current version is not functional and desperately needs a cleanup to make it so. I waffle between different implementations for similar behaviors and my organizations of objects and their methods is messy. In this regard, the weight of my inexperience is becoming burdensome. I'll probably reach out to a friend with more experience for advice on what to patch and which implementations are best practice.